postman collection is ready.

+++ src/main/java/com/akotako/loanservice/model/response/PayLoanResult.java
@@ -3,10 +3,12 @@ package com.akotako.loanservice.model.response;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class PayLoanResult {
     private int paidInstallmentsCount;
-    private double totalPaidAmount;
+    private BigDecimal totalPaidAmount;
     private boolean isLoanFullyPaid;
 }
